+++ charts/redpanda/chart_test.go
@@ -807,22 +807,22 @@ func mTLSValuesUsingCertManager() *redpanda.PartialValues {
 		External:      &redpanda.PartialExternalConfig{Enabled: ptr.To(false)},
 		ClusterDomain: ptr.To("cluster.local"),
 		Listeners: &redpanda.PartialListeners{
-			Admin: &redpanda.PartialAdminListeners{
+			Admin: &redpanda.PartialListenerConfig[redpanda.NoAuth]{
 				TLS: &redpanda.PartialInternalTLS{
 					RequireClientAuth: ptr.To(true),
 				},
 			},
-			HTTP: &redpanda.PartialHTTPListeners{
+			HTTP: &redpanda.PartialListenerConfig[redpanda.HTTPAuthenticationMethod]{
 				TLS: &redpanda.PartialInternalTLS{
 					RequireClientAuth: ptr.To(true),
 				},
 			},
-			Kafka: &redpanda.PartialKafkaListeners{
+			Kafka: &redpanda.PartialListenerConfig[redpanda.KafkaAuthenticationMethod]{
 				TLS: &redpanda.PartialInternalTLS{
 					RequireClientAuth: ptr.To(true),
 				},
 			},
-			SchemaRegistry: &redpanda.PartialSchemaRegistryListeners{
+			SchemaRegistry: &redpanda.PartialListenerConfig[redpanda.NoAuth]{
 				TLS: &redpanda.PartialInternalTLS{
 					RequireClientAuth: ptr.To(true),
 				},
@@ -856,7 +856,7 @@ func mTLSValuesWithProvidedCerts(serverTLSSecretName, clientTLSSecretName string
 			},
 		},
 		Listeners: &redpanda.PartialListeners{
-			Admin: &redpanda.PartialAdminListeners{
+			Admin: &redpanda.PartialListenerConfig[redpanda.NoAuth]{
 				//External: redpanda.PartialExternalListeners[redpanda.PartialAdminExternal]{
 				//	"default": redpanda.PartialAdminExternal{Enabled: ptr.To(false), Port: ptr.To(int32(0))},
 				//},
@@ -865,7 +865,7 @@ func mTLSValuesWithProvidedCerts(serverTLSSecretName, clientTLSSecretName string
 					Cert:              ptr.To("provided"),
 				},
 			},
-			HTTP: &redpanda.PartialHTTPListeners{
+			HTTP: &redpanda.PartialListenerConfig[redpanda.HTTPAuthenticationMethod]{
 				//External: redpanda.PartialExternalListeners[redpanda.PartialHTTPExternal]{
 				//	"default": redpanda.PartialHTTPExternal{Enabled: ptr.To(false), Port: ptr.To(int32(0))},
 				//},
@@ -874,7 +874,7 @@ func mTLSValuesWithProvidedCerts(serverTLSSecretName, clientTLSSecretName string
 					Cert:              ptr.To("provided"),
 				},
 			},
-			Kafka: &redpanda.PartialKafkaListeners{
+			Kafka: &redpanda.PartialListenerConfig[redpanda.KafkaAuthenticationMethod]{
 				//External: redpanda.PartialExternalListeners[redpanda.PartialKafkaExternal]{
 				//	"default": redpanda.PartialKafkaExternal{Enabled: ptr.To(false), Port: ptr.To(int32(0))},
 				//},
@@ -883,7 +883,7 @@ func mTLSValuesWithProvidedCerts(serverTLSSecretName, clientTLSSecretName string
 					Cert:              ptr.To("provided"),
 				},
 			},
-			SchemaRegistry: &redpanda.PartialSchemaRegistryListeners{
+			SchemaRegistry: &redpanda.PartialListenerConfig[redpanda.NoAuth]{
 				//External: redpanda.PartialExternalListeners[redpanda.PartialSchemaRegistryExternal]{
 				//	"default": redpanda.PartialSchemaRegistryExternal{Enabled: ptr.To(false), Port: ptr.To(int32(0))},
 				//},